Adding ssh-add to bashrc
After system reboot or login the ssh key is no longer saved in the
keystore. The first time trying to use the key requires a password the
use the key or the key has to be manually added to the keystore. This
change will check to see if the key exists on startup and add it if not.
Using a typical default location for the key so it can be shared across
systems.